📏 Architect: [Refactor/Health improvement] Split mcReducer.ts
Refactored mcReducer.ts to split the massive switch statement into smaller, domain-specific reducers (`bankReducer`, `missionReducer`, `responsibilityReducer`, `settingsReducer`, `gameReducer`, `systemReducer`). This drastically reduces the file size of `mcReducer.ts` and aligns with Single Responsibility Principle and Feature-Sliced Design to improve codebase maintainability.
