+++ src/mission-control/store/mcReducer.ts
@@ -18,13 +18,14 @@ import type {
 } from '../types';
 import { DEFAULT_SETTINGS } from '../types';
 
-function getLocalDateString(): string {
-    const d = new Date();
-    const year = d.getFullYear();
-    const month = String(d.getMonth() + 1).padStart(2, '0');
-    const day = String(d.getDate()).padStart(2, '0');
-    return `${year}-${month}-${day}`;
-}
+import {
+    bankReducer,
+    missionReducer,
+    responsibilityReducer,
+    settingsReducer,
+    gameReducer,
+    systemReducer
+} from './reducers';
 
 // ---- Default State ----
 
@@ -125,505 +126,25 @@ export const initialState: MCState = {
 // ---- Reducer ----
 
 function _mcReducer(state: MCState, action: MCAction): MCState {
-    switch (action.type) {
-        case 'ADD_TOKEN':
-            return { ...state, bankCount: state.bankCount + 1 };
+    const bankResult = bankReducer(state, action);
+    if (bankResult !== undefined) return bankResult;
 
-        case 'ADD_TOKENS':
-            return { ...state, bankCount: state.bankCount + action.amount };
+    const missionResult = missionReducer(state, action);
+    if (missionResult !== undefined) return missionResult;
 
-        case 'REMOVE_TOKEN':
-            return { ...state, bankCount: Math.max(0, state.bankCount - 1) };
+    const responsibilityResult = responsibilityReducer(state, action);
+    if (responsibilityResult !== undefined) return responsibilityResult;
 
-        case 'SELECT_CASE': {
-            const isQuickGame = action.reward === 'quick-game';
-            if (isQuickGame && state.gameTokens <= 0) return state;
+    const settingsResult = settingsReducer(state, action);
+    if (settingsResult !== undefined) return settingsResult;
 
-            return {
-                ...state,
-                gameTokens: isQuickGame ? state.gameTokens - 1 : state.gameTokens,
-                cases: state.cases.map(c =>
-                    c.id === action.caseId
-                        ? { ...c, status: 'active', reward: action.reward, targetCount: action.targetCount }
-                        : c,
-                ),
-            };
-        }
+    const gameResult = gameReducer(state, action);
+    if (gameResult !== undefined) return gameResult;
 
-        case 'DEPOSIT_TO_CASE': {
-            const amount = Math.min(action.amount, state.bankCount);
-            return {
-                ...state,
-                bankCount: state.bankCount - amount,
-                cases: state.cases.map(c =>
-                    c.id === action.caseId
-                        ? { ...c, tokenCount: c.tokenCount + amount }
-                        : c,
-                ),
-            };
-        }
+    const systemResult = systemReducer(state, action);
+    if (systemResult !== undefined) return systemResult;
 
-        case 'MOVE_TOKEN': {
-            // Source Validation
-            if (action.from === 'bank' && state.bankCount <= 0) return state;
-            if (typeof action.from === 'number') {
-                const sourceCase = state.cases.find(c => c.id === action.from);
-                if (!sourceCase || sourceCase.tokenCount <= 0) return state;
-            }
-
-            // Target Validation
-            if (typeof action.to === 'number') {
-                const targetCase = state.cases.find(c => c.id === action.to);
-                if (!targetCase || targetCase.status !== 'active') return state;
-                if (targetCase.reward === 'quick-game') return state;
-                if (targetCase.tokenCount >= targetCase.targetCount) return state;
-            }
-
-            let newBankCount = state.bankCount;
-            let newCases = state.cases;
-
-            // Decrement source
-            if (action.from === 'bank') {
-                newBankCount -= 1;
-            } else {
-                newCases = newCases.map(c => c.id === action.from ? { ...c, tokenCount: Math.max(0, c.tokenCount - 1) } : c);
-            }
-
-            // Increment target
-            if (action.to === 'bank') {
-                newBankCount += 1;
-            } else {
-                newCases = newCases.map(c => c.id === action.to ? { ...c, tokenCount: c.tokenCount + 1 } : c);
-            }
-
-            return {
-                ...state,
-                bankCount: newBankCount,
-                cases: newCases,
-            };
-        }
-
-        case 'VACUUM_TO_CASE': {
-            if (state.bankCount === 0) return state;
-            const vacuumTarget = state.cases.find(c => c.id === action.caseId);
-            if (!vacuumTarget) return state;
-            if (vacuumTarget.reward === 'quick-game') return state;
-            const canAdd = Math.min(state.bankCount, vacuumTarget.targetCount - vacuumTarget.tokenCount);
-            if (canAdd <= 0) return state;
-            return {
-                ...state,
-                bankCount: state.bankCount - canAdd,
-                cases: state.cases.map(c =>
-                    c.id === action.caseId
-                        ? { ...c, tokenCount: c.tokenCount + canAdd }
-                        : c,
-                ),
-            };
-        }
-
-        case 'REFUND_CASE': {
-            const targetCase = state.cases.find(c => c.id === action.caseId);
-            if (!targetCase) return state;
-            const isQuickGame = targetCase.reward === 'quick-game';
-            return {
-                ...state,
-                bankCount: state.bankCount + targetCase.tokenCount,
-                gameTokens: isQuickGame ? Math.min(5, state.gameTokens + 1) : state.gameTokens,
-                cases: state.cases.map(c =>
-                    c.id === action.caseId
-                        ? { ...c, status: 'empty', reward: null, tokenCount: 0 }
-                        : c,
-                ),
-            };
-        }
-
-        case 'SET_PRIVILEGE_STATUS':
-            return {
-                ...state,
-                privileges: state.privileges.map(p =>
-                    p.id === action.cardId
-                        ? { ...p, status: action.status, suspendedUntil: action.suspendedUntil }
-                        : p,
-                ),
-            };
-
-        case 'COMPLETE_TASK': {
-            const nextState = { ...state };
-            const currentMission = state.missions.find(m => m.phase === action.missionPhase);
-            const currentTask = currentMission?.tasks.find(t => t.id === action.taskId);
-            if (!currentTask) return state;
-
-            const wasCompleted = currentTask.completed;
-            const nextCompleted = !wasCompleted;
-
-            if (action.taskId === 'cream') {
-                const schedule = state.settings.creamTaskSchedule ?? 'evening';
-                const dec = schedule === 'both' ? 0.5 : 1;
-                if (nextCompleted) {
-                    nextState.creamTaskDaysLeft = Math.max(0, state.creamTaskDaysLeft - dec);
-                } else {
-                    nextState.creamTaskDaysLeft = state.creamTaskDaysLeft + dec;
-                }
-            }
-            nextState.missions = nextState.missions.map(m =>
-                m.phase === action.missionPhase
-                    ? {
-                        ...m,
-                        tasks: m.tasks.map(t =>
-                            t.id === action.taskId ? { ...t, completed: nextCompleted } : t,
-                        ),
-                    }
-                    : m,
-            );
-            // Auto-disable if today was the last day
-            if (action.taskId === 'cream' && nextState.creamTaskDaysLeft === 0) {
-                 nextState.settings = { ...nextState.settings, creamTaskEnabled: false };
-            }
-            return nextState;
-        }
-
-        case 'LOCK_TASK':
-            return {
-                ...state,
-                missions: state.missions.map(m =>
-                    m.phase === action.missionPhase
-                        ? {
-                            ...m,
-                            tasks: m.tasks.map(t =>
-                                t.id === action.taskId ? { ...t, locked: true } : t,
-                            ),
-                        }
-                        : m,
-                ),
-            };
-
-        case 'SET_ACTIVE_MISSION': {
-            // Timer expired — deactivate. Tasks are intentionally left as-is
-            // (they're reset when the mission re-triggers via SET_ACTIVE_MISSION).
-            if (action.phase === 'none') {
-                return {
-                    ...state,
-                    activeMission: 'none',
-                    missions: state.missions.map(m => ({
-                        ...m,
-                        active: false,
-                        durationMins: undefined,
-                    })),
-                };
-            }
-
-            // Only one mission at a time — if one is already running, ignore the new trigger
-            if (state.activeMission !== 'none') {
-                return state;
-            }
-
-            const now = new Date().toISOString();
-            return {
-                ...state,
-                activeMission: action.phase,
-                missions: state.missions.map(m => {
-                    if (m.phase !== action.phase) return { ...m, active: false };
-                    // Every trigger is a fresh start: new timer, reset checklist.
-                    const [sh, sm] = m.startsAt.split(':').map(Number);
-                    const [eh, em] = m.endsAt.split(':').map(Number);
-                    const durationMins = (eh * 60 + em) - (sh * 60 + sm);
-                    return {
-                        ...m,
-                        active: true,
-                        startedAt: now,
-                        durationMins: Math.max(0, durationMins), // no minimum — allows sub-minute test durations
-                        whiningDetected: false,
-                        whiningLocked: false,
-                        tasks: m.tasks.map(t => ({ ...t, completed: false, locked: false })),
-                    };
-                }),
-            };
-        }
-
-        // Reset task progress only — mission stays active, timer keeps running.
-        // Does NOT affect startedAt/durationMins/activeMission.
-        case 'RESET_MISSION':
-            return {
-                ...state,
-                missions: state.missions.map(m =>
-                    m.phase === action.missionPhase
-                        ? { ...m, active: true, loggedTimeoutAt: undefined, whiningDetected: false, whiningLocked: false, tasks: m.tasks.map(t => ({ ...t, completed: false, locked: false })) }
-                        : m
-                )
-            };
-
-        // Full reset — tasks AND timer restart from scratch.
-        // Mission stays active with a fresh startedAt + recalculated durationMins.
-        case 'RESET_MISSION_WITH_TIMER': {
-            const now = new Date().toISOString();
-            return {
-                ...state,
-                missions: state.missions.map(m => {
-                    if (m.phase !== action.missionPhase) return m;
-                    const [sh, sm] = m.startsAt.split(':').map(Number);
-                    const [eh, em] = m.endsAt.split(':').map(Number);
-                    let durationMins = (eh * 60 + em) - (sh * 60 + sm);
-                    if (durationMins < 0) durationMins += 24 * 60; // overnight wrap
-                    return {
-                        ...m,
-                        active: true,
-                        startedAt: now,
-                        durationMins,
-                        loggedTimeoutAt: undefined,
-                        whiningDetected: false,
-                        whiningLocked: false,
-                        tasks: m.tasks.map(t => ({ ...t, completed: false, locked: false })),
-                    };
-                }),
-            };
-        }
-
-        case 'CANCEL_MISSION':
-            return {
-                ...state,
-                activeMission: 'none',
-                missions: state.missions.map(m =>
-                    m.phase === action.missionPhase
-                        ? { ...m, startedAt: undefined, active: false, loggedTimeoutAt: undefined, whiningDetected: false, whiningLocked: false, tasks: m.tasks.map(t => ({ ...t, completed: false, locked: false })) }
-                        : m
-                )
-            };
-
-        case 'COMPLETE_MISSION_ROUTINE':
-            return {
-                ...state,
-                activeMission: 'none',
-                bankCount: state.bankCount + action.bonusTokens,
-                ...(action.missionPhase === 'morning' ? { lastCompletedOrFailedMorningDate: getLocalDateString() } : {}),
-                ...(action.missionPhase === 'evening' ? { lastCompletedOrFailedEveningDate: getLocalDateString() } : {}),
-                missions: state.missions.map(m =>
-                    m.phase === action.missionPhase
-                        ? { ...m, startedAt: undefined, active: false, loggedTimeoutAt: undefined, whiningDetected: false, whiningLocked: false, tasks: m.tasks.map(t => ({ ...t, completed: false, locked: false })) }
-                        : m
-                )
-            };
-
-        case 'MARK_MISSION_TIMEOUT':
-            return {
-                ...state,
-                ...(action.missionPhase === 'morning' ? { lastCompletedOrFailedMorningDate: getLocalDateString() } : {}),
-                ...(action.missionPhase === 'evening' ? { lastCompletedOrFailedEveningDate: getLocalDateString() } : {}),
-                missions: state.missions.map(m =>
-                    m.phase === action.missionPhase
-                        ? { ...m, loggedTimeoutAt: new Date().toISOString() }
-                        : m
-                )
-            };
-
-        case 'ADJUST_MISSION_END': {
-            return {
-                ...state,
-                missions: state.missions.map(m => {
-                    if (m.phase !== action.missionPhase) return m;
-                    // Ignore adjustments if mission is not actively running
-                    if (!m.active || !m.startedAt || m.durationMins == null) return m;
-                    
-                    return { ...m, durationMins: Math.max(1, m.durationMins + action.deltaMinutes) };
-                }),
-            };
-        }
-
-        case 'CONSUME_CASE': {
-            // Permanently remove tokens from case (reward redeemed) — NO bank refund
-            return {
-                ...state,
-                cases: state.cases.map(c =>
-                    c.id === action.caseId
-                        ? { ...c, status: 'empty' as const, reward: null, tokenCount: 0 }
-                        : c,
-                ),
-            };
-        }
-
-        case 'TOGGLE_WHINING':
-            return {
-                ...state,
-                missions: state.missions.map(m => {
-                    if (m.phase !== action.missionPhase) return m;
-                    if (m.whiningLocked && !action.lockedFromUI) return m; // UI clicks ignored if locked
-                    return {
-                        ...m,
-                        whiningDetected: !m.whiningDetected,
-                        whiningLocked: action.lockedFromUI ? true : m.whiningLocked
-                    };
-                }),
-            };
-
-        case 'SET_SETTINGS': {
-            const nextSettings = { ...state.settings, ...action.settings };
-            let nextDaysLeft = state.creamTaskDaysLeft;
-            
-            if (action.settings.creamTaskEnabled && !state.settings.creamTaskEnabled) {
-                // Just enabled — start fresh
-                nextDaysLeft = nextSettings.creamTaskDaysTarget;
-            } else if (action.settings.creamTaskDaysTarget !== undefined && action.settings.creamTaskDaysTarget !== state.settings.creamTaskDaysTarget) {
-                 // Target changed, reset current progress
-                 nextDaysLeft = action.settings.creamTaskDaysTarget;
-            }
-
-            // If the start time of the currently-active mission changes, we must also
-            // deactivate it — otherwise durationMins is wiped but active stays true,
-            // making the expiry check `durationMins != null` permanently false (hung mission).
-            const mornTimeChanged =
-                (action.settings.morningStartsAt ?? state.settings.morningStartsAt) !== state.settings.morningStartsAt;
-            const evenTimeChanged =
-                (action.settings.eveningStartsAt ?? state.settings.eveningStartsAt) !== state.settings.eveningStartsAt;
-            const activeIsBeingRescheduled =
-                (state.activeMission === 'morning' && mornTimeChanged) ||
-                (state.activeMission === 'evening' && evenTimeChanged);
-
-            return {
-                ...state,
-                settings: nextSettings,
-                creamTaskDaysLeft: nextDaysLeft,
-                // Deactivate root if the running mission's start time was changed
-                ...(activeIsBeingRescheduled ? { activeMission: 'none' as const } : {}),
-                // Live-update mission startsAt/endsAt from settings so scheduler picks them up.
-                // If the start time changes, clear startedAt, durationMins, and active
-                // so the scheduler can re-trigger at the new time without getting stuck.
-                missions: state.missions.map(m => {
-                    if (m.phase === 'morning') {
-                        const dur = action.settings.morningDurationMins ?? state.settings.morningDurationMins;
-                        const start = action.settings.morningStartsAt ?? state.settings.morningStartsAt;
-                        const [h, min] = start.split(':').map(Number);
-                        const endTotal = h * 60 + min + dur;
-                        const endsAt = `${String(Math.floor(endTotal / 60)).padStart(2, '0')}:${String(endTotal % 60).padStart(2, '0')}`;
-                        return {
-                            ...m,
-                            startsAt: start,
-                            endsAt,
-                            ...(mornTimeChanged ? { startedAt: undefined, durationMins: undefined, active: false } : {}),
-                        };
-                    }
-                    if (m.phase === 'evening') {
-                        const dur = action.settings.eveningDurationMins ?? state.settings.eveningDurationMins;
-                        const start = action.settings.eveningStartsAt ?? state.settings.eveningStartsAt;
-                        const [h, min] = start.split(':').map(Number);
-                        const endTotal = h * 60 + min + dur;
-                        const endsAt = `${String(Math.floor(endTotal / 60)).padStart(2, '0')}:${String(endTotal % 60).padStart(2, '0')}`;
-                        return {
-                            ...m,
-                            startsAt: start,
-                            endsAt,
-                            ...(evenTimeChanged ? { startedAt: undefined, durationMins: undefined, active: false } : {}),
-                        };
-                    }
-                    return m;
-                }),
-            };
-        }
-
-        case 'ADD_RESPONSIBILITY_POINT': {
-            const now = new Date().toISOString();
-            return {
-                ...state,
-                responsibilities: state.responsibilities.map(r => {
-                    if (r.id !== action.taskId) return r;
-                    if (r.completedAt && (action.amount || 1) > 0) return r; // already complete — ignore if adding
-                    const newPoints = Math.max(0, r.pointsEarned + (action.amount || 1));
-                    const isComplete = newPoints >= r.pointsRequired;
-                    return {
-                        ...r,
-                        pointsEarned: newPoints,
-                        completedAt: isComplete ? now : null,
-                    };
-                }),
-            };
-        }
-
-        case 'RESET_RESPONSIBILITY': {
-            const addedBank = action.claimTokens ? state.bankCount + action.claimTokens : state.bankCount;
-            return {
-                ...state,
-                bankCount: addedBank,
-                responsibilities: state.responsibilities.map(r =>
-                    r.id === action.taskId
-                        ? { ...r, pointsEarned: 0, completedAt: null }
-                        : r
-                )
-            };
-        }
-
-        case 'ADD_LOG': {
-            const newLogs = [action.log, ...(state.activityLogs || [])].slice(0, 200);
-            return {
-                ...state,
-                activityLogs: newLogs
-            };
-        }
-
-        case 'CLEAR_LOGS':
-            return {
-                ...state,
-                activityLogs: [],
-            };
-
-        case 'CHEAT_ATTEMPT':
-            return {
-                ...state,
-                hasUnreviewedCheatAttempt: true,
-            };
-
-        case 'CLEAR_CHEAT_FLAG':
-            return {
-                ...state,
-                hasUnreviewedCheatAttempt: false,
-            };
-
-        case 'GRANT_GAME_TOKEN': {
-            const todayStr = new Date().toISOString().slice(0, 10);
-            if (!action.force && state.gameTokensLastGrantedDate === todayStr) return state;
-            
-            const totalWealth = state.bankCount + state.cases.reduce((sum, c) => sum + c.tokenCount, 0);
-            if (totalWealth < 10 && !action.force) return state;
-            if (state.gameTokens >= 5) return state;
-            return {
-                ...state,
-                gameTokens: Math.min(5, state.gameTokens + 1),
-                gameTokensLastGrantedDate: action.force ? state.gameTokensLastGrantedDate : todayStr,
-            };
-        }
-
-        case 'CONSUME_GAME_TOKEN':
-            if (state.gameTokens <= 0) return state;
-            return { ...state, gameTokens: state.gameTokens - 1 };
-
-        case 'RESET_GAME_TOKENS':
-            return {
-                ...state,
-                gameTokens: 0,
-                gameTokensLastGrantedDate: null,
-            };
-
-        case 'TRIGGER_ANIMATION':
-            return {
-                ...state,
-                lastAnimationTrigger: {
-                    type: action.animation,
-                    timestamp: Date.now()
-                }
-            };
-
-        case 'START_GAME':
-            return {
-                ...state,
-                snakeGameActive: true
-            };
-
-        case 'END_GAME':
-            return {
-                ...state,
-                snakeGameActive: false
-            };
-
-        default:
-            return state;
-    }
+    return state;
 }
 
 // ---- Task Injection Sync ----
@@ -703,4 +224,3 @@ export function mcReducer(state: MCState, action: MCAction): MCState {
     }
     return nextState;
 }
-
