+++ src/mission-control/store/reducers/index.ts
@@ -1,0 +1,6 @@
+export { bankReducer } from './bankReducer';
+export { missionReducer } from './missionReducer';
+export { responsibilityReducer } from './responsibilityReducer';
+export { settingsReducer } from './settingsReducer';
+export { gameReducer } from './gameReducer';
+export { systemReducer } from './systemReducer';
